bugfix: forgot to account for write data synchronizer for branch table writes.
Was hidden since 8 threads running the same code get the correct entry
from the *previous* thread.
Also adjusted the initial thread so the initial read address is 0.
Fmax: 458/487/510
(avg. Fmax can be up to 1.6% lower over multiple sets of seed runs,
but that means an Fmax loss relative to previous from 0.4% to 2%,
a 5x difference!)