Don't clear CTL/DA/DB. Instead carry IOR/C around and use locally.
This approach saves some area (~33 registers and LUT)
and shows more clearly which signals wrap around the pipelines.
This is a bet that the local decode and write enable won't
become a critical path. However, that logic is very pipelined
and retime-able.
It will also make debugging easier as the CTL/DA/DB will be meaningful.