Remove A/B memory read address decoding. Re-do zero register.
Rather than decode 1-1023, decode 0-1024, and test for zero address
and use that to gate the read output. This means the RAM is active
when reading address 0, but now the decoding is pipelined, removing
a critical path to RAM read enable and conceptually redundant
read address decoding.
(The read operands are always in range of a data memory.)