Initial Opcode_Decoder_Memory.
Special case: no read address on names (write-only memory)
and the address of a name doesn't include the write_offset, so
it acts as the opcode number, since that's how the opcode is used
to address into that memory in the CPU, yielding the control bits
for that opcode.

Still set the write_offset at initialization so higher-level code
can generate run-time stores to that memory.