Go back to having labels at start of line. It's much nicer to look at.
Keep state in list of labels, opcodes, and branches.
If not found, it's a label. If found in any of the lists, it's an error.
Then strip the label from the line and recurse into line parser.
This doesn't require having labels on their own lines, but it will work.