Initial commit: Data commands parser and front end.
We can now build symbolic descriptions of variables, pointers, etc...
in the data front-end.
This allows us to fill in metadata as we go along and go back to update
things like which memory a variable will reside in, as the code is
parsed. Once ready, we can call the back end to generate the binary
image data.