Fix Usage: allocate Branch Counter, Sentinel, and Mask at the same index as the allocated Branch Detector
Otherwise, when we reach a branch, the necessary configuration is in another
Branch Detector which isn't firing. This change is necessary because
the init loads are allocated when the branches are first encountered,
so the allocation of BDs was out of sync with that of BCs, S, and Ms.

Code tested in simulation, and looks OK, but not confirmed yet.