Change control interfaces to enable/done, which allows a clearer sequencer.
The control interface is level-based, as that allows a plain
handshake to enable and return a done signal.
The system interface takes pulse inputs, but outputs levels,
since when a CPU does a read/write it is only sampling/inserting values,
and we have no idea of the timing, only the sequence.